[#164] Refactored Graph Component Closes #164
* Added Graph Container
* Updated paths and replaced creation of GraphComponent with GraphContainer
